Lots of changes and refactoring on the Amf Module.
- BinaryStream was cleaned up and got some unittests.
- there are now distinct functions for all the unsigned types.
- switching between big and little endian reading/writing is implemented

- All the type mapping is  done by a class implementing Zend_Amf_TypeMapperInterface.
- There is full support for ArrayCollection and ArrayList. These are
implemented as ArrayObjects on the php side.
- There is a default implementation of the externalizable feature by
implementing Zend_Amf_Util_ExternalizableInterface
- As the type mapper is exchangeable, there is no need for the user to
depend on Zend_Amf_Util_ExternalizableInterface. Everyone can choose
its own schema. Even the _explicitType and getASClassName functionality can be
arbitrarily modified by writing a custom TypeMapper

There are some backwards incompatibilities though:
- Zend_Amf_Parse_TypeLoader is no more. If it is really needed, we could
reimplement it and modify the default TypeMapper in a way, that the values
from the Zend_Amf_Parse_TypeLoader are used. (the old functionality is still there. e.g.
Zend_Amf_Parse_TypeLoader::setResourceLoader is now
$server->getTypeMapper()->setResourceLoader() , in case the default
TypeMapper is used)
- BinaryStream::readInt() now reads 32bit values. The old readInt
functionality is now in readUnsignedShort()
- BinaryStream::readByte() now reads a signed byte. The old readByte is now
in readUnsignedByte()
- Zend_Amf_Request doesn't do the parsing in initialize() instead
there is a new parse() method. This is needed, as the request doesn't know the
TypeMapper at the time initialize() is called.